requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
langgraph==0.2.0
langchain==0.2.0
langchain-google-genai==1.0.6
apscheduler==3.10.4
gspread==6.1.2
google-auth==2.29.0
chromadb==0.5.0
python-dotenv==1.0.1
requests==2.32.3
pytz==2024.1
pydantic==2.7.1
groq==0.9.0